Refactor storybook example for map validation
